+++ src/main/java/roomescape/presentation/controller/AdminController.java
@@ -2,16 +2,15 @@ package roomescape.presentation.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.business.service.ReservationService;
 import roomescape.presentation.dto.ReservationRequest;
 import roomescape.presentation.dto.ReservationResponse;
 
-@Controller
+@RestController
 @RequestMapping("/admin")
 public class AdminController {
 
@@ -19,26 +18,6 @@ public class AdminController {
 
     public AdminController(final ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public String home() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservations() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String theme() {
-        return "admin/theme";
     }
 
     @PostMapping("/reservations")
