+++ src/main/java/roomescape/presentation/controller/page/AdminPageController.java
@@ -1,0 +1,30 @@
+package roomescape.presentation.controller.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminPageController {
+
+    @GetMapping
+    public String index() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservations() {
+        return "admin/reservation-new";
+    }
+
+    @GetMapping("/time")
+    public String time() {
+        return "admin/time";
+    }
+
+    @GetMapping("/theme")
+    public String theme() {
+        return "admin/theme";
+    }
+}
