+++ src/main/java/roomescape/presentation/controller/page/UserPageController.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.controller;
+package roomescape.presentation.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,15 +6,20 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping
-public class UserController {
+public class UserPageController {
 
     @GetMapping("/reservation")
     public String reservation() {
         return "reservation";
     }
 
-    @GetMapping(value = "reservation-mine")
+    @GetMapping("reservation-mine")
     public String reservationMine() {
         return "reservation-mine";
+    }
+
+    @GetMapping("/login")
+    public String login() {
+        return "login";
     }
 }
